Fixed crash of PETSc interface tests in CMake build
I don't fully understand what the issue was, but it appears the
constructor/destructor pair was not being called properly for
PetscLib. This resulted in PETSc not being properly initialised for
later tests. Whatever the exact details, the problem arose in
IndexerTest/TestGetRegionBndry and seems to have been related to a
for-loop with the iterator being a reference to a GlobalIndexer
object. Switching to using pointers got rid of the problem.