Renamed test-laplace3d to test-laplace-petsc3d
This name better reflects what the test is for. It is not for testing
a truly 3D Laplace solver (a non-working version of which used to be
includd in the code). Instead it is for an implementation of the 2D
solver (using PETSc) which may include some y-derivative terms due to
a curved metric.