Add y-communications in Hermite spline interpolations
The HermiteSpline, MonotonicHermiteSpline and HermiteSplineOnlyZ
interpolation classes were relying on the old behaviour of
Mesh::communicateXZ(), which despite the name actually communicated also
in y, and just did not calculate the parallel slices. To regain this
behaviour, use Mesh::sendY() and Mesh::wait() directly.

Corner cells of fx, fz, and fxz, which may actually be used, are now
communicated correctly.