Ensure coordinates can be constructed
interp_to, communicate() and to/fromFieldAligned require that the
coordinates have been constructed.
* Thus some care has to be taken, to e.g. only call communicateXZ rather
then communicate. Otherwise communicate might also calculate the
parallel fields, which fails because coordinate is not valid.
* While calling interp_to, we need to ensure fields are fieldaligned if
the direction is Y.
* the free to/from fieldaligned function read the fields coordinate
system. While in the constructor, this hasn't been yet set. But we
already now it, so we can directly call it, if we have created the
parallel transform.
* create3D tries to interpolate from fieldAligned. This doesn't work in
the constructor, and we need to do this manually (done in
interpolateAndExtrapolate) with maybeFromFieldAligned