Use separate pvode_initialised flag for PvodeSolver
Solver::init() sets Solver::initialised=true. The destructor
PvodeSolver::~PvodeSolver() previously used initialised to check if it
should free memory. Solver::init() is called near the start of
PvodeSolver::init(), so Solver::initialised is set to true before
PvodeSolver has completed initialisation. An error during the
initialisation (which is not unlikely, as CvodeMalloc calls the rhs
function!) could then cause a segfault as the destructor free'd memory
that had not been fully allocated. Using a separate pvode_initialised
flag, which is set to true only at the end of PvodeSolver::init() fixes
this issue.