Fix shift-transform
The shift-transform in staggered location requieres dz.
To calculate the staggered dz in y direction, the parallel transform is
requiered. Thus we special handel the case if dz is constant, and we can
do a trivial staggering, without a parallel transform.
For non-const dz, staggered dz needs to be given in an input file, so we
don't have to interpolate.

Also make isConst a free function, similar to min and max.