Attempt to update staggered grid support in reconnect-2field example
This is a best-effort update, and may not work correctly. By default the
option mesh:StaggerGrids=false for reconnect-2field, and staggered grid
support should be regarded as experimental. In particular, BOUT++'s
Grad_parP operator does not support staggered grids, and so the
implementation here for a modified Grad_parP of calling the built-in
Grad_parP and then interpolating the result may not be a very good
choice.