Move checks for J and Bxy out of jacobian() method
Perform checks in Coordinates constructors after all initialisation of J
and Bxy. Calculating J and Bxy from the metric coefficients may
sometimes cause problems with positivity (which would previously have
caused an exception in the jacobian() method), which can be worked
around by loading a consistent J and Bxy from the grid file. So it makes
sense to perform the checks only after trying to load J and Bxy from the
grid.