Set number of smoother steps for AMG test
Fixes #2089

This is necessary due to a change in PETSc 3.11:

> Previously the KSP Chebyshev implementation always did one more
> iteration than requested. For example -ksp_max_it 2 resulted in 3
> Chebyshev iterations. This has been corrected. Due to this correction
> your solver may seem to converge more slowely than it previous
> has. Note that the multigrid solvers (PCMG, PCGAMG, PCML) used by
> default Chebyshev (with 3 actual steps) in their smoother, thus the
> multigrid solvers will now have seemingly different convergence rates
> since they will now use only 2 actual steps. To reproduce previous
> behavior change the number of smoother iterations to match the
> previous actual amount, this can be done with for example
> -mg_levels_ksp_max_it 3 (or -prefix_mg_levels_ksp_max_it 3 if the KSP
> object has a prefix).

This was also the cause of #2034, which was fixed in #2051

We should be able to set the default value with `PCMGSetNumberSmooth`,
but this requires us to call `PCMGSetLevels` first, but we don't have
the information to do that when setting up the solver. We also can't
set `mg_levels_ksp_max_it` as a PETSc option in the solver, because
that needs to go to the `PetscLib` instance.

Therefore, we need to either set `mg_levels_ksp_max_it` in the tests,
or adjust the tolerance for the tests.