Don't read/write FieldPerps in guard cells
To be more precise, mark the global yindex for FieldPerps in guard
cells as invalid when writing them, and set the (local) yindex to be
invalid if its in a guard cell on this processor when reading them.

It is still possible to write FieldPerps in guard cells, it's just
that `collect` won't read them, and the user will need to use a
lower-level API to do so.