Generate random run ID, track restarts
Improve data provenance tracking, so that sets of runs can be linked
to each other, and the run used to generate data identified.

Each time the solver is run, generate a random number which is
broadcast to all processors. This is stored as "run_id" in:
- log files
- BOUT.settings, though at the moment only if the run finishes
- restart files
- dump files

When a simulation is restarted, the ID of the run it started from is
also recorded as "run_restart_from".

If restart files are used which don't have a run_id, then this will
either cause the run to fail, or set run_id to 0. The
`run_restart_from` ID therefore has two special values:
- 1 means no restart, the run was started from scratch
- 0 means restart from unknown run_id (missing -> run_id set to 0)

There may well be a better way to handle missing run_id values in
restart files.