Tidy up coordinate systems with explicit directions
Define the directions of the poloidal and toroidal angles more
carefully, so that at each step the direction of each coordinate is
clearer.

The new calculations clarify what happens when the poloidal field is
negative. For the standard coordinate system (where J can be >0 or <0),
the metric tensor doesn't change from what it was before. The
calculation of the field line pitch and integrated shear however has
been clarified (in terms of which toroidal direction is used).

A new coordinate system is defined, with the Y coordinate reversed for
negative poloidal field. This always has a positive Jacobian, and some
metric elements reverse sign, but has the same properties (being Clebsch).