Remove derivative workarounds in Coordinates
All derivatives used while creating Coordinates objects are in the
geometry() method. There was a problem that standard versions of the
derivatives need a Coordinates pointer in the Fields passed in.  By
moving the call to geometry() out of the constructor, we have removed
the need for workarounds to take derivatives before there is a valid
Coordinates pointer.