Check for Field2D in interp_to, don't shift to/from field-aligned
Field2Ds are axisymmetric and don't need to shift to or from
field-aligned. With the recent change removing the special
toFieldAligned override in field2d.hxx, instead relying on the templates
in field.hxx and the ParallelTransform::toFieldAligned(Field2D) and
ParallelTransform::fromFieldAligned(Field2D) methods, interp_to would
fail on a Field2D when constructing a ShiftedMetric transform by
interpolating zShift. Skipping the to/from field-aligned transformations
entirely for Field2D avoids this.