Exception creating ShiftedMetric by interpolating with 3d metrics
When using 3d metrics it is not possible to construct a ShiftedMetric
transform by interpolating zShift to CELL_YLOW, because the
interpolation would require a shift from field-aligned coordinates using
the ShiftedMetric which is being constructed (since zShift is not
axisymmetric when 3d metrics are allowed). Now throw an exception in
this case, and suggest creating zShift_CELL_YLOW in the grid file as a
solution.