Bugfix: Don't clobber the OptionsNetCDF file on second call to write
Creating an `OptionsNetCDF` with `FileMode::replace` overwrites the
file on _every_ call to `OptionsNetCDF::write`, which is probably not
what you want. One way around this would be to have the `NcFile`
object as a member variable and keep it open between calls to `write`.
This has the complication of us then needing to make sure to flush the
file correctly.

Instead, just change the mode after writing to file.