Bugfix: Constructing Options from brace init-list missed full_name
Constructing `Options` using an initializer_list like:

    Options option {{"section1", {"value1", 42}}};

would not initialise the section names. This mostly down to the value
being constructed from the very bottom of the tree. When we construct
the bottom-most value, we don't know its parent's name, so we can't
give it its full_name "section1:value1".

Instead, we need to do this recursively, as we construct each parent.