Use operator+= instead of getUnion() in ZInterpolation
getUnion() was misleadingly named, and actually gave the intersection.
Region::operator+=(Region) updates a Region to be the union of two
Regions, which is almost what was wanted, although we have to call
mask_region.unique() afterwards too to ensure there are no repeated
entries and the entries are sorted.