Update CMakeLists.txt
examples/laplacexy/simple doesn't compile under Cuda. In the very short term just to test this PR it is commented out in Top level CMakeLists.txt.  This app needs changes to its CMakeLists.txt or should not be compiled under Cuda - both are very easy fixes. It will be remedied in next update to next-hypre-outerloop-cuda