Restrict region for toFieldAligned() in VDDY() when not staggered
When the derivative is not staggered, vel_aligned does not require
y-guard cells, so toFieldAligned() can use "RGN_NOBNDRY" instead of
"RGN_NOX". This is a slight optimisation, and avoids an exception that
can be thrown from b0xGrad_dot_Grad() when using ShiftedMetricInterp,
which is thrown because a result is calculated in uninitialised guard
cells where it is not needed.