Fix memory leak of Coordinates in unit tests
The Coordinates' metric elements can hold shared_ptrs that result in a
cyclic dependency between the Fields and the Coordinates, with the
result that some unit tests have memory leaks.

One place the issue can arise is when multiplying fields by metric
elements with `CHECK > 0`. Then, `ASSERT1_FIELDS_COMPATIBLE` checks
that both fields share a `Coordinates` -- but this has the side effect
of initialising the `shared_ptr` in the metric elements.

This means that when the `Mesh` gets destroyed, the `coords_map`
starts to destroy its contents, but because the `Coordinates` metric
elements members have references to the `Coordinates`, the
`Coordinates` in `coords_map` doesn't get destroyed. The end result is
that the `Coordinates` gets leaked.

The fix here is to use a `weak_ptr` in `Field`. This only gets a
reference to the `Coordinates` when needed.