Add an option to CyclicReduce to gather onto fewer processors
By default all processors are used, resulting in an all-to-all
communication pattern.

An optional third argument to the constructor now specifies how many
processors should be used. This reduces the number of messages, at the
cost of increasing load imbalance. The processors used are evenly
distributed, in the hope that this will minimise network traffic on
any given node.