Add ngather option to Laplace cyclic solver
Input option to control the number of processors (in X) to gather
systems onto. This can be used to tune performance: Smaller ngather
leads to fewer messages, but more load imbalance.

Modified test, to cover more variations of nxpe and ngather.

That test found some bugs, which are fixed here.