Mark possible boundary condition inputs as used
We now (soft) error if there are any unused input parameters, but
because the set of boundary regions can be different on different
processors, we might not "use" (that is, read the value of) the
corresponding input parameters on the other processors!

This function is a bit of a workaround for the above: essentially
marks all possible boundary condition inputs as having been used on
each processor. This way, we should still catch invalid inputs and
misspellings (for example, `bdry_oll`) while still allowing valid
boundaries on other processors.