Rewrite outerloop version of elm-pb
Merging operations into a single large loop, which can then
be iterated over using RAJA or BOUT_FOR.

To enable operations to be merged, run-time switches have to be
converted to compile-time:

- Compile-time flags (e.g. NONLINEAR) are set to `true` or `false` at
  the top of the `elm-pb-outerloop.cxx` file.
- At run-time the options are read in the same way as `elm-pb`
- The run-time value is compared against the compile-time value
  (e.g. input `nonlinear` and compile flag `NONLINEAR`). If they
  are different then an exception is thrown. This is so that the
  simulation always uses the value expected from the input file.
  For convenience a `CHECK_SETTING` macro is defined.
- To turn on and off terms in the equations, an `EVAL_IF` macro
  is defined to improve readability.

Terms which have been moved to single index operators are left in
the code for now, but commented out. This is to allow comparison
of results while testing.