Starting new CoordinatesAccessor design
If too many CoordinatesFieldAccessors are added to FieldAccessor,
then the CUDA kernel parameter list runs out of space.
-> Need to reduce the number of pointers stored in FieldAccessor

This leads to having a struct to hold the Coordinates pointers,
which is then shared between the FieldAccessors.

That struct must be in memory accessible to the GPU, so either
allocated using Umpire or a local variable which can be captured by
value by the RAJA lambda function. The local variable would need
to be created by the user and passed into FieldAccessor constructors
or as an extra argument to operators.

This design instead uses a single BoutReal array, allocated in Array
using Umpire, to store all the Coordinates data. The different fields
(dx,dy,dz etc) are copied into the array and striped, so that values
for a single grid cell are next to each other in memory. Hopefully
this will help with memory access performance.

To avoid doing this copy for every FieldAccessor, the arrays are
cached in a map from Coordinates* to Array<BoutReal>. Unfortunately
if Coordinates members are changed, this change will not be updated
in the CoordinatesAccessor unless this store is cleared.