Move RAJA wrappers to bout/rajalib.hxx
BOUT_FOR_RAJA can be used with or without RAJA; if RAJA is not
available (or DISABLE_RAJA is true) then it falls back to using
BOUT_FOR.

RajaForAll struct which wraps RAJA::forall only available if RAJA
is available. Could implement a fall-back which uses BOUT_FOR
if it turns out to be useful.