Adding build-all-examples target
CMake switch BOUT_BUILD_EXAMPLES now controls whether the examples
directory is added. It is ON by default, and adds the examples
with the EXCLUDE_FROM_ALL flag.

To actually build all examples, this adds a build-all-examples target,
and examples are added to it. Specific examples can be built by using
specific targets or cd'ing into their directory (in the build directory)
and running "make".

If BOUT_BUILD_EXAMPLES is turned OFF, then the examples are not copied
into the build directory. This could be used if BOUT++ is a sub-module
of another project.