Make cmake submodule update work with older versions of git
In older versions of git `submodule update --init --recursive` had to be
run from the root of the repo. This commit replaces
`CMAKE_CURRENT_SOURCE_DIR` with `PROJECT_SOURCE_DIR` in the cmake call
to update the submodules to ensure that the update runs from the repo
root, making it compatible with older git versions.