Workaround for conflicts between different libuuid instances
Sometimes (e.g. when using conda) there can be multiple copies of
libuuid.so in LD_LIBRARY_PATH. This commit provides a new CMake variable
`SYMLINK_SYSTEM_UUID` - if this is set to ON then symlink the found
system header/library into the build directory's include/ and lib/
subdirectories.

By default, CMake adds an 'rpath' for linked libraries (at least when
just building and not 'installing') so that at run time the system
linker finds the exact same version that CMake configured with. However,
as I understand it this cannot work with system libraries, e.g. in
/usr/, because the rpath is only to a directory to search, not to the
particular library, so it does not make sense to add generic system
libraries (not sure how this is avoided, but it seems to be). 'rpath' is
great because it avoids relying on LD_LIBRARY_PATH (which then has to be
set up correctly at run-time), but for system libraries this breaks
down. CMake does know (during the configure steps) what exact files it
should link to though, so if we symlink those into the build directory
(CMAKE_CURRENT_BINARY_DIR) then we can add the build directory to the
rpath and have it take precedence over everything in LD_LIBRARY_PATH.
Then we avoid any potential conflicts with different versions of
libraries in LD_LIBRARY_PATH (that might be introduced by conda, etc.).