Option for CMake to ignore anything installed by conda
If BOUT_IGNORE_CONDA_ENV=ON, then ignore anything under $CONDA_PREFIX.
This means that even if there is an active conda environment, no
conda-installed packages will be detected/used/linked.

Implementation is hack-y because CMake does not currently have a way to
exclude a prefix, so we have to make a list including every
sub-directory. This may be easier in future, see
https://gitlab.kitware.com/cmake/cmake/-/merge_requests/6880.