Using python based on version is a pain.
I have newer python versions installed for testing, but they dont have all
the libraries installed, thus bout++ always fails. If a new enough cmake is
present we should find the one that appears first in PATH, rather then the
most recent one. It is easy to change PATH to ensure the correct python is found.