Update Solver::iteration from BoutMonitor::call()
Make `Solver::iteration` private, and add methods to get or increment
its value.

Allows the iteration counter to be updated directly in
`BoutMonitor::call()`.

Previously, `Solver::iteration` was updated in the implementations of
`Solver::run()`, but if there is any monitor that is called more
frequently than the output monitor, this meant that `iteration` counted
the number of calls to the more frequent monitor, not the number of
output steps. Not an issue now, since it is incremented in
`BoutMonitor`, which is the one that does the output.