Move update of Solver::iteration into SolverMonitor
Previously, `Solver::iteration` was updated in the implementations of
`Solver::run()`, but if there is any monitor that is called more
frequently than the output monitor, this means that `iteration` counts
the number of calls to the more frequent monitor, not the number of
output steps.

Now, have a `SolverMonitor` in the base `Solver` class, that is called
at the output frequency, and updates the iteration count.