Fix off-by-1 in saved `iteration` variable
Previously, the `iteration` saved into dump files was inconsistent with
`hist_hi`: in a simulation that is not restarted (so starts from
iteration 0), `iteration` would always be 1 less than `hist_hi`.
`hist_hi` counts the number of output steps that have been completed,
which seems more sensible, so this commit adds 1 to the value of
`iteration` that is saved. Note this also makes the value that is saved
the same as the value that is printed to stdout at the same time.