Handle periodicX in cyclic Laplace solver
If `periodicX = true` then the (kx,kz) = (0,0) mode is unconstrained
and the Laplacian matrix inversion of vorticity to get potential
becomes singular.

The solution is to force solutions to have a zero average potential.
This fix retains the tridiagonal form of the matrix by treating the
kz=0 mode specially:
- In the tridiagonal solve fix one location (x=xstart) to zero
- After the solve, calculate and subtract the X average of the kz=0
  modes i.e. set the (0,0) mode to zero.

For the Hasegawa-wakatani model this results in solutions which are
periodic in X, have zonal (kz=0) modes, and run at least as quickly as
the non-periodic simulations.