Make iteration more robust and give more options in LaplaceNaulin
Previously a cycle decreasing the underrelax_factor was triggered if the
error increased on an iteration. However sometimes it can be more robust
to just continue the iteration with moderate increases in the error -
there are cases when decreasing the underrelax_factor never results in a
decreasing error, but continuing the main iteration loop for a few steps
would. This commit tries to handle this kind of case, with several new
parameters which can be adjusted to control how and when
underrelax_factor is decreased.