ADIOS2 and array index mapping
In general file I/O maps a logically rectangular subset of
the local (nx, ny, nz) array onto a part of a global array.
That global array may be sparse i.e. there might be holes in it.

The mapping does not cover the entire local array because we
insert communication guard cells around the outside.

To perform this mapping we use ADIOS Put() function to get
a span of a buffer, and then iterate over fields to copy data
into those buffers.