Use generic Factory for OptionsIO
Sorry for the large commit; this touches quite a few places.

The OptionsNetCDF and OptionsADIOS classes are now private,
so that OptionsIO is now the interface to both.

The generic Factory is used to implement OptionsIOFactory,
and the "netcdf" and "adios" types are registered in their
separate header files.

The OptionsIOFactory handles the choice of default section
("restart_files" for restarts, "output" for dump/output files),
and sets the default path and prefix. This simplifies both
the PhysicsModel and the separate implementations by putting
that logic in one place.