ADIOS reading of groups, more test fixes
ADIOS uses '/' as a separator, but BOUT++ options uses ':'.
When reading variables, we now split on '/' separator and navigate
to the correct subsection. Attributes are also named using the final
part of the name after the last '/'.

solver and beuler integrated tests weren't calling BoutInitialise,
so failed when ADIOSFinalize is called.