Changing Options string to bool behaviour
Enables boolean options expressions like:
```
thing = true
another = false
third = thing & another
```

but breaks inputs using "yes", "False" and other strings that are not
"true" or "false" for boolean settings.

When converting a string to a BoutReal, Options uses an expression
parser that can use variables defined in the input options. The same
parser is used when converting to an integer. When converting to
a bool, however, Options was doing a simple case-insensitive match to
"true", "false", "yes", "no" and a few similar strings.

This commit changes this behaviour, treating bools like a special
case of an integer: The same parser is used, but the result must
be either 1 or 0. This enables boolean quantities to behave consistently
with int and BoutReal types.

We now define "true" to be equal to 1, and "false" to be 0, so
can be used in expressions e.g `true * 2 -> 2`, and `false * 3 -> 0`.
No other variations so e.g. "yes", "no", "True" are undefined.

We define operators `|` and `&` for logical `or`, and logical
`and`. Due to the simplicity of the parser binary operators must be
single characters.  These both check that their left and right
arguments are bools (i.e. either 1 or 0). Logical not (`!`) is not
implemented yet, and may require changes to the parser (I guess treat
like a unary `-`).