Options: Delete copy constructor and copy assignment
Accidentally copying Options leads to unexpected bugs where
changes (e.g documentation) changes a copy rather than the
original. Usually what is intended is a reference, but the
author omits an `&`.

Copy assignment is also ambiguous: Is it the value, the
attributes, or the children that should be copied?

Now if a copy is really intended, there is a `.copy()` method
(thanks Rust!)