Use a wrapper function to handle FormFunctionForColoring signature mismatch
Using reinterpret_cast no longer works.

Expected signature is PetscErrorCode (*)()
but is PetscErrorCode (*)(SNES, Vec, Vec, void*) on main branch.

For the newer version the current solver instance must be passed as the 'cxt' argument,
so store 'this' instance in a global variable (before calling MatFDColoringSetFunction).

Have to do it this way because the function passed as the (second) argument to MatFDColoringSetFunction must be a free function,
so cannot make the function wrapper an instance method and pass 'this' from there.