tests: Fix for 3D metric in FCI test
The Div_par operators use parallel slices of B -- with 2D metrics,
these are just the field itself, in 3D we need the actual
slices. While `Coordinates::geometry` does communicate the fields, it
puts off calculating the parallel slices because that needs the fully
constructed `Coordinates`.

Upcoming changes should fix this and remove the need to explicitly
communicate `Coordinates` members.