snes solver: Consider timestep change at output
Previously the timestep adjustment was inside a conditional
`if (looping) {...`. The variable `looping` is false when the
simulation time reaches or exceeds the next output time.

This meant that when the internal timestep exceeded the output
step, `looping` was almost always `false` and so the internal
timestep would saturate.

With this change a timestep adjustment will be considered after
every successful internal timestep.

Note: There may have been a reason why timestep adjustments were
not considered, but I can't remember what it was and it was not
documented.