snes solver: Pseudo Transient Continuation method
Setting `pseudo_time = true` now enables Pseudo-timestepping,
in which each cell has a separate timestep.

The timestep in each cell is set inversely proportional to the
residual (RMS time-derivative of all quantities in cell).

Recommend enabling `pid_controller` that multiplies all timesteps
by the same factor, to control the number of nonlinear iterations.

Output time uses the minimum timestep in the domain, so simulations
shouldn't need to run for as long as when all cells are evolved at
the same speed.